+++ protocols/autonat/examples/server.rs
@@ -26,6 +26,7 @@
 //! ```
 //! The `listen-port` parameter is optional and allows to set a fixed port at which the local peer should listen.
 
+use clap::Parser;
 use futures::prelude::*;
 use libp2p::autonat;
 use libp2p::identify::{Identify, IdentifyConfig, IdentifyEvent};
@@ -34,12 +35,11 @@ use libp2p::swarm::{Swarm, SwarmEvent};
 use libp2p::{identity, Multiaddr, NetworkBehaviour, PeerId};
 use std::error::Error;
 use std::net::Ipv4Addr;
-use structopt::StructOpt;
 
-#[derive(Debug, StructOpt)]
-#[structopt(name = "libp2p autonat")]
+#[derive(Debug, Parser)]
+#[clap(name = "libp2p autonat")]
 struct Opt {
-    #[structopt(long)]
+    #[clap(long)]
     listen_port: Option<u16>,
 }
 
@@ -47,7 +47,7 @@ struct Opt {
 async fn main() -> Result<(), Box<dyn Error>> {
     env_logger::init();
 
-    let opt = Opt::from_args();
+    let opt = Opt::parse();
 
     let local_key = identity::Keypair::generate_ed25519();
     let local_peer_id = PeerId::from(local_key.public());