+++ protocols/mdns/CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.36.0 [unreleased]
+# 0.37.0 [unreleased]
+
+- Update to `libp2p-core` `v0.33.0`.
+
+- Update to `libp2p-swarm` `v0.36.0`.
+
+# 0.36.0
 
 - Update to `libp2p-swarm` `v0.35.0`.